update: handles case of island on border of grid
One of the conditions states that the grid is surrounded by water(0), so
it is assumed that land should not consume the borders of the grid.
Should this case occur, the island cells(1) on the border of the grid
shall be ignored to the extent of not counting as a cell until referred
to by a different cell.