bug: skipped island perimeter checks
When checking an island cell's perimeter with an index suroassing the range of the grid, the IndexError exception caught by the try block causes the code to 'pass' and thus skip any of the other perimeter checks that the if statements were intended to make.

The bug was solved by removing the try and accounting for every possible IndexError for the island cell being checked. All the nested statements may look a bit messy, but it solves the problem for now.