add: one solution to the prime game challenge
This solution makes of the Sieve of Eratosthenes algorithm to efficiently find prime numbers. The function assumes the numbers will only go no higher than 10 000.
Instead of removing all the multiples of the prime number a player choses, the winner is determined by the number of prime numbers found. Maria always wins when with an odd number of prime numbers found. Ben wins when an even number of prime numbers are found.